Move v1 api tests
